flask/requirements.txt
Flask
pymongo[srv]
Flask-PyMongo
flask_mongoengine
flask_mysqldb